refactor: 기능별 model 분리
