+++ gpt/models/mindmap_response.py
@@ -1,5 +1,6 @@
+from typing import List
+
 from pydantic import BaseModel, Field
-from typing import List, Literal
 
 
 class GeneratedQuestionsResponse(BaseModel):
@@ -18,9 +19,3 @@ class SummarizedNodeResponse(BaseModel):
     summary: str = Field(
         ..., description="질문과 답변을 요약한 결과"
     )
-
-
-class EisenhowerRecommendationResponse(BaseModel):
-    recommendedQuadrant: Literal["Q1", "Q2", "Q3", "Q4"] = Field(..., description="추천된 아이젠하워 사분면")
-    reason: str = Field(..., description="해당 사분면이 추천된 이유")
-    isSameAsCurrent: bool = Field(..., description="현재 사분면과 동일한지 여부")
